ddcontroller: Add support for PWM/DIR BLDC motors
Available BLDC motors with built-in drivers can be controlled with
a single PWM signal and a GPIO as a direction control (CW/CCW). This
simplifies the PWM control, and requires the use of a hardware PWM
as the PWM frequency required by the motors is typically 10 kHz to
30 kHz. This has the added advantage of greatly reducing noise in
audible range.

This patch adds a 'control_type' parameter to the motor type.
Set this to 'PWM_DIR' to use the new mode, or 'DUAL_PWM' (default)
to utilize a DC motor controlled by 2 PWMs and driven by an
h-bridge.

In 'PWM_DIR' mode, the first control signal number is the PWM, and
the second is the direction control signal.

In addition, there is a setting for max and min duty cycle. After
experimentation, it was found that these motors have a linear duty
cycle range that results in rotation and that this range can start
well above 10% (around 40%) and can peak before 100% (around 95%).